Change to the current version of VBISAM available
This change pushes the vbisam library updates, and also some changes
made to the CFFI backend compilation script.

[Issue 1]

+++ vbisam2/libvbisam/isdecimal.c
@@ -431,7 +431,7 @@ dectoflt (dec_t *dp, float *fltp)
 	int status;
 
 	status = dectoreal (dp, &dbl, 8);
-	*fltp = dbl;
+	*fltp = (float)(dbl);
 	return status;
 }
